Make it work without APP_NAME again
Everything is using the search store for now.
