--- src/core/client.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { render } from 'react-dom';
-import { Router, browserHistory } from 'react-router';
-import routes from './routes';
-
-render(
-  <Router children={routes} history={browserHistory} />,
-  document.getElementById('react-view')
-);
