+++ src/core/client/index.js
@@ -1,0 +1,5 @@
+import makeClient from './base';
+import routes from './routes';
+import createStore from 'search/store';
+
+makeClient(routes, createStore);
