+++ src/core/routes.js
@@ -1,9 +1,14 @@
 import React from 'react';
-import { Route } from 'react-router';
+import { IndexRoute, Route } from 'react-router';
 
+import App from './containers/App';
 import discoRoutes from '../disco/routes';
 import searchRoutes from '../search/routes';
 
 export default (
-  <Route children={[discoRoutes, searchRoutes]} />
+  <Route path="/">
+    <IndexRoute component={App} />
+    {searchRoutes}
+    {discoRoutes}
+  </Route>
 );
